+++ database/seeders/DatabaseSeeder.php
@@ -4,6 +4,7 @@ namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class DatabaseSeeder extends Seeder
 {
@@ -14,7 +15,7 @@ class DatabaseSeeder extends Seeder
     {
         DB::table('admins')->insert([
             'email' => 'admin@example.com',
-            'password' => 'adminadmin'
+            'password' => Hash::make('adminadmin'),
         ]);
     }