+++ routes/web.php
@@ -5,7 +5,6 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\RegistrationFormController;
 use App\Http\Controllers\ConferenceListController;
 use App\Http\Controllers\AdminAccessController;
-use App\Models\User;
 
 /*
  Routes: registration form 
@@ -20,15 +19,16 @@ Route::view('/edit_form', 'registrationPartials/register');
  Routes: conference list 
 */
 Route::get('/conference_list', [ConferenceListController::class, 'index']);
+Route::get('/table_body', [ConferenceListController::class, 'table']);
 
 /*
  Routes: admin access
 */
 Route::view('/sign_in', 'sign-in')->middleware('guest');
-Route::view('/sign_out', 'sign-out')->middleware(['auth', 'verified']);
+Route::view('/sign_out', 'sign-out')->middleware(['auth']);
 
-Route::post('/get_admin_access', [AdminAccessController::class, 'signIn']);
-Route::post('/leave_admin_access', [AdminAccessController::class, 'signOut']);
+Route::post('/get_admin_access', [AdminAccessController::class, 'signIn'])->middleware('guest');
+Route::post('/leave_admin_access', [AdminAccessController::class, 'signOut'])->middleware(['auth']);
 
 Route::view('/add_speaker', 'adminPartials/add');
 Route::get('/edit_speaker/{id}', [ConferenceListController::class, 'edit']);