+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-- Created 'Sign In' page view with success flash message. 
+# [0.2.0] - 2023-03-21 
+
+### Fixed 
+
+- Updated table style in 'Conference List' view.  
+
+### Added 
+
+- Created 'Sign In', 'Sign Out' page views with success flash message. 
 - Added DB seeder with admin access ['email' => 'admin@example.com', 'password' => 'adminadmin'].
-- Updated table style in 'Conference List' view and added admin functions.  
+- Added admin functions. 
 
 # [0.1.1] - 2023-03-15